coordinator/requirements.txt
coloredlogs
pyyaml
pymongo
requests
